+++ libs/journeys/ui/src/libs/getStepHeading/getStepHeading.spec.ts
@@ -1,7 +1,17 @@
 import { TreeBlock } from '..'
+import { BlockFields_StepBlock as StepBlock } from '../transformer/__generated__/BlockFields'
 import { getStepHeading } from '.'
 
 describe('getStepHeading', () => {
+  const stepBlock: StepBlock = {
+    __typename: 'StepBlock',
+    id: 'step.id',
+    parentBlockId: null,
+    parentOrder: null,
+    locked: false,
+    nextBlockId: null
+  }
+
   it('returns text of first typography block', () => {
     const children: TreeBlock[] = [
       {
@@ -41,11 +51,43 @@ describe('getStepHeading', () => {
       }
     ]
 
-    expect(getStepHeading(children)).toEqual('Heading')
+    const treeBlocks: TreeBlock[] = [
+      {
+        ...stepBlock,
+        children
+      }
+    ]
+
+    expect(
+      getStepHeading({ blockId: 'step.id', children, treeBlocks })
+    ).toEqual('Heading')
   })
 
-  it('returns undefined if there are no typography blocks', () => {
+  it('returns step number if there are no typography blocks', () => {
     const children: TreeBlock[] = []
-    expect(getStepHeading(children)).toBeUndefined()
+    const treeBlocks: TreeBlock[] = [
+      {
+        ...stepBlock,
+        children
+      }
+    ]
+
+    expect(
+      getStepHeading({ blockId: 'step.id', children, treeBlocks })
+    ).toEqual('Step 1')
+  })
+
+  it('returns Untitled step if no typogrpahy blocks and id not matched', () => {
+    const children: TreeBlock[] = []
+    const treeBlocks: TreeBlock[] = [
+      {
+        ...stepBlock,
+        children
+      }
+    ]
+
+    expect(
+      getStepHeading({ blockId: 'anotherStep.id', children, treeBlocks })
+    ).toEqual('Untitled step')
   })
 })
